Methods for DeeDeeExperiment objects

The DeeDeeExperiment() class provides a family of methods to get and set DE-related information and functional enrichment results in DeeDeeExperiment objects.

# S4 method for class 'DeeDeeExperiment'
dea_info(x)

# S4 method for class 'DeeDeeExperiment'
dea_info(x) <- value

# S4 method for class 'DeeDeeExperiment'
dea_names(x)

# S4 method for class 'DeeDeeExperiment'
add_dea(x, dea)

# S4 method for class 'DeeDeeExperiment,character'
remove_dea(x, dea_name)

# S4 method for class 'DeeDeeExperiment,character'
dea(x, dea_name, verbose = TRUE)

# S4 method for class 'DeeDeeExperiment'
get_dea_list(x, verbose = TRUE)

# S4 method for class 'DeeDeeExperiment'
fea_info(x)

# S4 method for class 'DeeDeeExperiment'
fea_info(x) <- value

# S4 method for class 'DeeDeeExperiment'
fea_names(x)

# S4 method for class 'DeeDeeExperiment'
add_fea(x, fea_res, de_name = NA_character_, fe_name = NULL)

# S4 method for class 'DeeDeeExperiment,character'
remove_fea(x, fea_name)

# S4 method for class 'DeeDeeExperiment,character'
fea(x, fea_name, verbose = TRUE)

Arguments

x

A DeeDeeExperiment() object

value

Replacement value for replacement methods.

dea

A named list of DE results, in any of the formats supported by the package (currently: results from DESeq2, edgeR, limma).

dea_name

Character value, specifying the name of the DE analysis to get or remove

verbose

Logical, whether or not to display warnings. If TRUE, warnings/messages will be displayed. If FALSE, the function runs silently

fea_res

A data frame or a named list of data frames containing fea results. Each element should represent a table of enrichment terms (e.g., GO terms).

de_name

A character string to explicitly specify the name of the de result this fea should be linked to. If not provided, the function will attempt to match fea names to de results automatically.

fe_name

A character string giving a name to the FE results. It can only support FDR to adjust the threshold used for summarizing DE results.

fea_name

Character value, specifying the name of the functional enrichment result to add or remove

fea

A named list of Functional Enrichment results

Value

Return value varies depending on the individual methods, as described below.

Details

dea_info and dea_info<- are the methods to get and set the dea information as a whole. These methods return DeeDeeExperiment objects.

add_dea and remove_dea are used to respectively add or remove DE-results items. These methods also return DeeDeeExperiment objects, with updated content in the dea slot.

dea_names returns the names of the available DE contrasts in DeeDeeExperiment objects.

dea and get_dea_list retrieve the dea information and provide this as a DataFrame object (for a specific analysis) or as a list, with one element for each reported analysis.

fea and fea<- are the methods to get and set the fea information as a whole. These methods return DeeDeeExperiment objects.

fea_names returns the names of the available enrichment results in DeeDeeExperiment objects.

add_fea and remove_fea are used to respectively add or remove functional enrichment results items. These methods also return DeeDeeExperiment objects, with updated content in the fea slot.

show is the method to nicely print out the information of a DeeDeeExperiment object.

Examples

data("de_named_list", package = "DeeDeeExperiment")
library("SummarizedExperiment")
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> Warning: package ‘matrixStats’ was built under R version 4.6.0
#> 
#> Attaching package: ‘MatrixGenerics’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> Loading required package: GenomicRanges
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> Loading required package: generics
#> Warning: package ‘generics’ was built under R version 4.6.0
#> 
#> Attaching package: ‘generics’
#> The following objects are masked from ‘package:base’:
#> 
#>     as.difftime, as.factor, as.ordered, intersect, is.element, setdiff,
#>     setequal, union
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, is.unsorted, lapply,
#>     mapply, match, mget, order, paste, pmax, pmax.int, pmin, pmin.int,
#>     rank, rbind, rownames, sapply, saveRDS, table, tapply, unique,
#>     unsplit, which.max, which.min
#> Loading required package: S4Vectors
#> 
#> Attaching package: ‘S4Vectors’
#> The following object is masked from ‘package:utils’:
#> 
#>     findMatches
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: GenomeInfoDb
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: ‘Biobase’
#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians

rd_macrophage <- DataFrame(
  gene_id = rownames(de_named_list$ifng_vs_naive))
rownames(rd_macrophage) <- rownames(de_named_list$ifng_vs_naive)
se_macrophage_noassays <- SummarizedExperiment(
  assays = SimpleList(),
  rowData = rd_macrophage
)

# creating a `DeeDeeExperiment`
dde <- DeeDeeExperiment(
  se_macrophage_noassays,
  de_results = de_named_list
)
dde
#> class: DeeDeeExperiment 
#> dim: 17806 0 
#> metadata(1): version
#> assays(0):
#> rownames(17806): ENSG00000000003 ENSG00000000419 ... ENSG00000285982
#>   ENSG00000285994
#> rowData names(13): gene_id ifng_vs_naive_log2FoldChange ...
#>   salmo_both_pvalue salmo_both_padj
#> colnames: NULL
#> colData names(0):
#> dea(4): ifng_vs_naive, ifngsalmo_vs_naive, salmonella_vs_naive, salmo_both 
#> fea(0): 

new_del <- list(
  ifng2 = de_named_list$ifng_vs_naive,
  ifngsalmo2 = de_named_list$ifngsalmo_vs_naive
)

# add a new (set of) DE result(s)
dde_new <- add_dea(dde, new_del)
dde_new
#> class: DeeDeeExperiment 
#> dim: 17806 0 
#> metadata(1): version
#> assays(0):
#> rownames(17806): ENSG00000000003 ENSG00000000419 ... ENSG00000285982
#>   ENSG00000285994
#> rowData names(19): gene_id ifng_vs_naive_log2FoldChange ...
#>   ifngsalmo2_pvalue ifngsalmo2_padj
#> colnames: NULL
#> colData names(0):
#> dea(6): ifng_vs_naive, ifngsalmo_vs_naive, salmonella_vs_naive, salmo_both, ifng2, ifngsalmo2 
#> fea(0): 

# removing DEAs
dde_removed <- remove_dea(dde, "ifng_vs_naive")
dde_removed
#> class: DeeDeeExperiment 
#> dim: 17806 0 
#> metadata(1): version
#> assays(0):
#> rownames(17806): ENSG00000000003 ENSG00000000419 ... ENSG00000285982
#>   ENSG00000285994
#> rowData names(10): gene_id ifngsalmo_vs_naive_log2FoldChange ...
#>   salmo_both_pvalue salmo_both_padj
#> colnames: NULL
#> colData names(0):
#> dea(3): ifngsalmo_vs_naive, salmonella_vs_naive, salmo_both 
#> fea(0):